+++ ErsatzTV/Services/RunOnce/ResourceExtractorService.cs
@@ -61,6 +61,12 @@ public class ResourceExtractorService : BackgroundService
             FileSystemLayout.ChannelGuideTemplatesFolder,
             stoppingToken);
 
+        await ExtractTemplateResource(
+            assembly,
+            "_musicVideo.sbntxt",
+            FileSystemLayout.ChannelGuideTemplatesFolder,
+            stoppingToken);
+
         await ExtractScriptResource(
             assembly,
             "_threePartEpisodes.js",